move ingest ascwds to new folder structure
