+++ _01_ingest/ascwds/tests/test_ingest_ascwds_dataset.py
@@ -1,13 +1,15 @@
 import unittest
 from unittest.mock import call, patch, Mock
 
-import jobs.ingest_ascwds_dataset as job
+import _01_ingest.ascwds.jobs.ingest_ascwds_dataset as job
+from _01_ingest.unittest_data.data import IngestASCWDSData as Data
+from _01_ingest.unittest_data.schemas import IngestASCWDSData as Schemas
 from utils.column_names.raw_data_files.ascwds_worker_columns import (
     AscwdsWorkerColumns as AWK,
 )
-from tests.test_file_data import IngestASCWDSData as Data
-from tests.test_file_schemas import IngestASCWDSData as Schemas
 from utils import utils
+
+PATCH_PATH = "_01_ingest.ascwds.jobs.ingest_ascwds_dataset"
 
 
 class IngestASCWDSDatasetTests(unittest.TestCase):
@@ -25,8 +27,8 @@ class IngestSingleFileTest(IngestASCWDSDatasetTests):
     def setUp(self) -> None:
         super().setUp()
 
-    @patch("utils.utils.construct_destination_path")
-    @patch("jobs.ingest_ascwds_dataset.handle_job")
+    @patch(f"{PATCH_PATH}.ingest_utils.construct_destination_path")
+    @patch(f"{PATCH_PATH}.handle_job")
     def test_ingest_single_file(
         self, handle_job_mock: Mock, construct_destination_path_mock: Mock
     ):
@@ -54,10 +56,10 @@ class IngestSingleFileTest(IngestASCWDSDatasetTests):
 
 
 class IngestMultipleFilesTests(IngestASCWDSDatasetTests):
-    @patch("utils.utils.construct_s3_uri")
-    @patch("utils.utils.construct_destination_path")
-    @patch("utils.utils.get_s3_objects_list")
-    @patch("jobs.ingest_ascwds_dataset.handle_job")
+    @patch(f"{PATCH_PATH}.ingest_utils.construct_s3_uri")
+    @patch(f"{PATCH_PATH}.ingest_utils.construct_destination_path")
+    @patch(f"{PATCH_PATH}.ingest_utils.get_s3_objects_list")
+    @patch(f"{PATCH_PATH}.handle_job")
     def test_ingest_multiple_files(
         self,
         handle_job_mock: Mock,
@@ -115,12 +117,12 @@ class IngestMultipleFilesTests(IngestASCWDSDatasetTests):
 
 
 class TestHandleJob(IngestASCWDSDatasetTests):
-    @patch("utils.utils.read_partial_csv_content")
-    @patch("utils.utils.identify_csv_delimiter")
-    @patch("utils.utils.read_csv")
-    @patch("utils.utils.write_to_parquet")
-    @patch("jobs.ingest_ascwds_dataset.raise_error_if_mainjrid_includes_unknown_values")
-    @patch("jobs.ingest_ascwds_dataset.fix_nmdssc_dates")
+    @patch(f"{PATCH_PATH}.ingest_utils.read_partial_csv_content")
+    @patch(f"{PATCH_PATH}.ingest_utils.identify_csv_delimiter")
+    @patch(f"{PATCH_PATH}.ingest_utils.read_csv")
+    @patch(f"{PATCH_PATH}.utils.write_to_parquet")
+    @patch(f"{PATCH_PATH}.raise_error_if_mainjrid_includes_unknown_values")
+    @patch(f"{PATCH_PATH}.fix_nmdssc_dates")
     def test_handle_job_calls_correct_functions_when_dataset_is_ascwds(
         self,
         fix_nmdssc_dates_mock: Mock,
@@ -156,12 +158,12 @@ class TestHandleJob(IngestASCWDSDatasetTests):
         fix_nmdssc_dates_mock.assert_not_called()
         write_to_parquet_mock.assert_called_once_with(df, self.destination_path)
 
-    @patch("utils.utils.read_partial_csv_content")
-    @patch("utils.utils.identify_csv_delimiter")
-    @patch("utils.utils.read_csv")
-    @patch("utils.utils.write_to_parquet")
-    @patch("jobs.ingest_ascwds_dataset.raise_error_if_mainjrid_includes_unknown_values")
-    @patch("jobs.ingest_ascwds_dataset.fix_nmdssc_dates")
+    @patch(f"{PATCH_PATH}.ingest_utils.read_partial_csv_content")
+    @patch(f"{PATCH_PATH}.ingest_utils.identify_csv_delimiter")
+    @patch(f"{PATCH_PATH}.ingest_utils.read_csv")
+    @patch(f"{PATCH_PATH}.utils.write_to_parquet")
+    @patch(f"{PATCH_PATH}.raise_error_if_mainjrid_includes_unknown_values")
+    @patch(f"{PATCH_PATH}.fix_nmdssc_dates")
     def test_handle_job_calls_correct_functions_when_dataset_is_nmdssc(
         self,
         fix_nmdssc_dates_mock: Mock,
@@ -198,10 +200,10 @@ class TestHandleJob(IngestASCWDSDatasetTests):
         raise_error_mock.assert_not_called()
         write_to_parquet_mock.assert_called_once_with(df, self.destination_path)
 
-    @patch("utils.utils.read_partial_csv_content")
-    @patch("utils.utils.identify_csv_delimiter")
-    @patch("utils.utils.read_csv")
-    @patch("utils.utils.write_to_parquet")
+    @patch(f"{PATCH_PATH}.ingest_utils.read_partial_csv_content")
+    @patch(f"{PATCH_PATH}.ingest_utils.identify_csv_delimiter")
+    @patch(f"{PATCH_PATH}.ingest_utils.read_csv")
+    @patch(f"{PATCH_PATH}.utils.write_to_parquet")
     def test_handle_job_raises_error_if_dataset_is_not_a_valid_option(
         self,
         write_to_parquet_mock: Mock,
