+++ terraform/pipeline/glue.tf
@@ -82,7 +82,7 @@ module "clean_cqc_pir_data_job" {
 
 module "ingest_ascwds_dataset_job" {
   source          = "../modules/glue-job"
-  script_dir      = "jobs"
+  script_dir      = "_01_ingest/ascwds/jobs"
   script_name     = "ingest_ascwds_dataset.py"
   glue_role       = aws_iam_role.sfc_glue_service_iam_role
   resource_bucket = module.pipeline_resources
